# payment-service/api/openapi.yaml
openapi: 3.0.1
info:
  title: Payment API
  version: 1.0.0

servers:
  - url: http://localhost:8080/v1

paths:
  /api/payments/initiate:
    post:
      summary: Initiates a payment
      tags:
        - payment-controller
      operationId: initiatePayment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InitiatePaymentRequest'
      responses:
        200:
          description: Payment initiated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InitiatePaymentResponse'

  /api/payments/verify/{transactionId}:
    get:
      summary: Verifies a payment
      tags:
        - payment-controller
      operationId: verifyPayment
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: string
            example: "80bbc31b-2989-4b18-b873-c7348f37a6da"
      responses:
        200:
          description: Payment verification details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyPaymentResponse'

  /api/payments:
    get:
      summary: fetches user payments
      tags:
        - payment-controller
      operationId: getUserPayments
      parameters:
        - name: userId
          in: query
          required: false
          schema:
            type: string
            example: "bc66c48a67fc-1722290520376"
      responses:
        200:
          description: User payments details
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GetUserPaymentsResponse'

components:
  schemas:
    InitiatePaymentRequest:
      type: object
      description: Request payload for initiating a payment transaction.
      properties:
        customer:
          $ref: "#/components/schemas/CustomerDetails"
        product:
          $ref: "#/components/schemas/ProductDetails"
        payment:
          $ref: "#/components/schemas/PaymentInfo"
      required:
        - user
        - product
        - payment

    CustomerDetails:
      type: object
      description: User (customer) information required for the transaction.
      properties:
        id:
          type: string
          description: Unique identifier for the user.
          example: "683b04c9b2ea75795fc091e0"
      required:
        - id

    ProductDetails:
      type: object
      description: Product details being purchased.
      properties:
        name:
          type: string
          description: Name of the product or service.
          example: "PLAN_50"
        quantity:
          type: integer
          format: int32
          description: Number of items being purchased.
          example: 1
      required:
        - name
        - quantity

    PaymentInfo:
      type: object
      description: Payment details for the transaction.
      properties:
        amount:
          type: number
          format: double
          description: Transaction amount.
          example: 56.75
        currency:
          type: string
          description: Three-letter ISO 4217 currency code.
          example: "GBP"
          enum:
            - USD
            - GBP
            - EUR
            - JPY
      required:
        - amount
        - currency

    InitiatePaymentResponse:
      type: object
      properties:
        transactionId:
          type: string
          example: "80bbc31b-2989-4b18-b873-c7348f37a6da"
        paymentLink:
          type: string
          example: "https://localhost.com:8090"

    VerifyPaymentResponse:
      type: object
      properties:
        paymentStatus:
          type: string
          example: "SUCCESS"
        referenceId:
          type: string
          example: "REF123456789"

    GetUserPaymentsResponse:
      type: object
      properties:
        transactionId:
          type: string
          example: "d85ed2bd-e59d-425b-8ac5-1330a7290ec9-1743986763"